COOKEEP-61 프로필 이미지 조회 API를 별도 컨트롤러로 분리
